RsaKeys v0.1.0 RsaKeys

Documentation for RsaKeys.

Functions

decrypt(data, priv_key, password)

Nif function for decrypt data with private key and password

decrypt_data(data, password, priv_key \\ privkey())

Specs

decrypt_data(Binary.t, String.t, Binary.t) :: Binary.t

Decrypt data with encrypted data, password, private key and save to file

encrypt(data, key)

Nif function for encrypt data with public key

encrypt_data(data, pub_key \\ pubkey())

Specs

encrypt_data(Binary.t, Binary.t) :: Binary.t

Encrypt data with data, public key and save to file

generate_keys(password)

Specs

generate_keys(String.t) :: {atom, Binary.t, Binary.t}

generate_keys(String // password)

## Examples

#

iex> RsaKeys.generate_keys(“123123”)

{:ok,

“——-BEGIN RSA PRIVATE KEY——-

Proc-Type: 4,ENCRYPTED

DEK-Info: DES-EDE3-CBC,3123FE7FDE3D2D06

#

vUI5BtcWw88KE0q…….

——-END RSA PRIVATE KEY——-

“,

“——-BEGIN PUBLIC KEY——-

MIIBIjANB…….

——-END PUBLIC KEY——-

“}

load_nif()

Loading nif functions

privkey()

Specs

privkey :: Binary.t

Get private key from file

pubkey()

Specs

pubkey :: Binary.t

Get public key from file

rsagen()

Nif function for generate rsa keys with des3

save_to_file(data, filename)

Specs

save_to_file(Binary.t, String.t) :: Binary.t

Save binary data to priv folder with filename